requirements.txt
torch
torchvision
pytorch-lightning
opencv-python-headless
numpy
pyyaml
tqdm
pillow
matplotlib
scikit-learn
timm